+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0
+- Moved: controllers to http directory
+- Updated: minor changes to payment controller to be more concise
+
 ## 0.1.0
 - fix psr-4 naming problem
 - minor template fixes